fix some warnings
git-svn-id: file:///.data/svn_convert/svn@6523 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ branches/branch-QString/plugins/netmonitor/netmonitor.cpp
@@ -137,7 +137,7 @@ void NetmonitorPlugin::setLogType(unsigned id, bool bLog)
     }
 }
 
-const unsigned NO_DATA = (unsigned)(-1);
+const long NO_DATA = -1;
 
 void NetmonitorPlugin::showMonitor()
 {
